textbook_ingestion/requirements.txt
# GCS
google-cloud-storage==2.18.2

# File extraction (DOCX only — PDFs use Vision OCR directly)
python-docx==1.1.2

# OCR (Google Vision — processes PDFs natively, no Poppler needed)
google-cloud-vision==3.7.4

# Neo4j
neo4j==5.26.0

# Gemini LLM via Vertex AI (new google-genai SDK)
google-genai>=1.0.0

# Shared
python-dotenv==1.0.1
